[refactor] 파일 구조 변경 : repository 를 domain 에 합침     - api.repository.BoardRepository => api.domain.BoardRepository     - api.repository.MemberRepository => api.domain.MemberRepository     - 나머지는 import 경로 수정.
