+++ src/test/java/com/inhabas/api/board/repository/JpaBoardRepositoryTest.java
@@ -4,8 +4,8 @@ import com.inhabas.api.domain.board.NormalBoard;
 import com.inhabas.api.domain.board.Category;
 import com.inhabas.api.domain.member.IbasInformation;
 import com.inhabas.api.domain.member.Member;
-import com.inhabas.api.repository.board.BoardRepository;
-import com.inhabas.api.repository.member.MemberRepository;
+import com.inhabas.api.domain.board.BoardRepository;
+import com.inhabas.api.domain.member.MemberRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
