+++ src/test/java/com/inhabas/api/member/MemberServiceTest.java
@@ -4,20 +4,16 @@ import com.inhabas.api.domain.member.IbasInformation;
 import com.inhabas.api.domain.member.Member;
 import com.inhabas.api.domain.member.Role;
 import com.inhabas.api.domain.member.SchoolInformation;
-import com.inhabas.api.repository.member.MemberRepository;
+import com.inhabas.api.domain.member.MemberRepository;
 import com.inhabas.api.service.member.MemberService;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityExistsException;
-import java.util.Date;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 @SpringBootTest
